feat(league): implement league and league team subcommands
