+++ Cargo.toml
@@ -11,7 +11,7 @@ path = "src/main.rs"
 
 [dependencies]
 clap = { version = "4.5.23", features = ["derive"] }
-fbsim-core = "1.0.0-alpha.13"
+fbsim-core = { git = "https://github.com/whatsacomputertho/fbsim-core.git", branch = "league-schedule-gen" }
 indicatif = "0.17.11"
 rand = "0.8.5"
 serde_json = "1.0.134"
