+++ src/game.rs
@@ -1,0 +1,2 @@
+pub mod benchmark;
+pub mod sim;
